+++ src/main/java/jenkins/branch/OrganizationFolderViewHolder.java
@@ -189,13 +189,7 @@ public class OrganizationFolderViewHolder extends AbstractFolderViewHolder {
      * A custom category specific view.
      */
     @Restricted(NoExternalUse.class)
-    public static class ViewImpl extends ListView {
-
-        /**
-         * The category that this view selects.
-         */
-        @NonNull
-        private final SCMSourceCategory category;
+    public static class ViewImpl extends BaseView<SCMSourceCategory> {
 
         /**
          * Creates a new view.
@@ -204,10 +198,9 @@ public class OrganizationFolderViewHolder extends AbstractFolderViewHolder {
          * @param category the category.
          */
         public ViewImpl(ViewGroup owner, @NonNull SCMSourceCategory category) {
-            super(category.getName(), owner);
-            this.category = category;
+            super(owner,category);
             try {
-                getJobFilters().replaceBy(asList(new MultiBranchCategoryFilter(category)));
+                getJobFilters().replaceBy(List.of(new MultiBranchCategoryFilter(category)));
                 getColumns().replaceBy(asList(
                         new StatusColumn(),
                         new WeatherColumn(),
@@ -218,50 +211,6 @@ public class OrganizationFolderViewHolder extends AbstractFolderViewHolder {
                 // ignore
             }
             setRecurse(false);
-        }
-
-        /**
-         * {@inheritDoc}
-         */
-        @Override
-        public String getDisplayName() {
-            return category.getDisplayName() + " (" + getItems().size() + ")";
-        }
-
-        /**
-         * {@inheritDoc}
-         */
-        @Override
-        public boolean isRecurse() {
-            return false;
-        }
-
-        /**
-         * {@inheritDoc}
-         */
-        @NonNull
-        @Override
-        public ACL getACL() {
-            final ACL acl = super.getACL();
-            return new ACL() {
-                @Override
-                public boolean hasPermission2(@NonNull Authentication a, @NonNull Permission permission) {
-                    if (View.CREATE.equals(permission)
-                            || View.CONFIGURE.equals(permission)
-                            || View.DELETE.equals(permission)) {
-                        return false;
-                    }
-                    return acl.hasPermission2(a, permission);
-                }
-            };
-        }
-
-        /**
-         * {@inheritDoc}
-         */
-        @Override
-        public void save() throws IOException {
-            // no-op
         }
 
         /**
